Modify install.sh to be silent

+++ install.sh
@@ -1,15 +1,22 @@
-mkdir data
-mkdir results
+mkdir -p data
+mkdir -p results
+
 cd results
-mkdir 01-make-manifest
-mkdir 02-make-artifact
-mkdir 03-infer-asv
-mkdir 04-merge-rep-seqs
-mkdir 05-make-phylogeny
-mkdir 06-merge-table
-mkdir 07-make-rarefaction-curve
-mkdir 08-export-rarefaction-curve
-mkdir 09-calculate-plateau
-mkdir 10-rarefy
-ln -s 06* 11-merge-rarefied-table
+
+mkdir -p 01-make-manifest
+mkdir -p 02-make-artifact
+mkdir -p 03-infer-asv
+mkdir -p 04-merge-rep-seqs
+mkdir -p 05-make-phylogeny
+mkdir -p 06-merge-table
+mkdir -p 07-make-rarefaction-curve
+mkdir -p 08-export-rarefaction-curve
+mkdir -p 09-calculate-plateau
+mkdir -p 10-rarefy
+
+if ! [ -L 11-merge-rarefied-table ] ; then
+    ln -s 06* 11-merge-rarefied-table
+fi
+
+mkdir -p 12-filter-rep-seqs
 cd ..
